backend/requirements.txt
fastapi
uvicorn
numpy
pytesseract
opencv-python-headless
pillow
gtts
groq
